계정관리 수정 기능 추가

+++ app/controllers/User.go
@@ -62,6 +62,51 @@ func (u User) SelectUserList() revel.Result {
 	return u.RenderJSON(items)
 }
 
+func (u User) LoadUserList() revel.Result {
+	//session check
+	result := make(map[string]interface{})
+	authKey, ok := u.Session["authKey"]
+	if !ok || authKey == "" {
+		u.Response.Status = http.StatusUnauthorized
+		return u.RenderJSON(result)
+	}
+
+	var reqParam = make(map[string]interface{})
+	u.Params.BindJSON(&reqParam)
+
+	id := reqParam["id"]
+	revel.INFO.Println("LoadUserList id...====", id)
+	DB := dbConn()
+	rows, err := DB.QueryContext(ctx, "SELECT id,password,role,username FROM thx_employee WHERE id= ?", id)
+	if err != nil {
+		log.Fatal(err)
+	}
+	defer rows.Close()
+
+	var items []userInfo
+
+	for rows.Next() {
+		item := userInfo{}
+
+		err := rows.Scan(
+			&item.Id,
+			&item.Password,
+			&item.Role,
+			&item.Username,
+		)
+		CheckErr(err)
+
+		items = append(items, item)
+		return u.RenderJSON(item)
+	}
+	//revel.INFO.Println("items====", items)
+
+	if err != nil {
+		revel.INFO.Println("DB Error", err)
+	}
+	return u.RenderJSON(items)
+}
+
 func (u User) AddUserSave() revel.Result {
 	result := make(map[string]interface{})
 	authKey, ok := u.Session["authKey"]
@@ -82,26 +127,30 @@ func (u User) AddUserSave() revel.Result {
 	//revel.INFO.Println("AddUserSave role...====", role)
 	//revel.INFO.Println("AddUserSave username...====", username)
 
-	DB := dbConn()
-	// Insert
-	stmt, err := DB.Prepare("INSERT thx_employee SET id=?,password=?, role=?, username=?")
-	CheckErr(err)
+	if id != nil && username != nil && password != nil && role != nil {
+		DB := dbConn()
+		// Insert
+		stmt, err := DB.Prepare("INSERT thx_employee SET id=?,password=?, role=?, username=?")
+		CheckErr(err)
 
-	res, err := stmt.Exec(id, password, role, username)
-	CheckErr(err)
+		res, err := stmt.Exec(id, password, role, username)
+		CheckErr(err)
 
-	//// sql.Result.RowsAffected() 체크
-	n, err := res.RowsAffected()
-	if n == 1 {
-		revel.INFO.Println("1 row inserted.")
-	}
-	CheckErr(err)
+		//// sql.Result.RowsAffected() 체크
+		n, err := res.RowsAffected()
+		if n == 1 {
+			revel.INFO.Println("1 row inserted.")
+		}
+		CheckErr(err)
 
-	if err != nil {
+		if err != nil {
+			result["auth"] = "fail"
+			revel.INFO.Println("DB Error", err)
+		}
+		result["auth"] = "success"
+	} else {
 		result["auth"] = "fail"
-		revel.INFO.Println("DB Error", err)
 	}
-	result["auth"] = "success"
 
 	return u.RenderJSON(result)
 }
@@ -120,21 +169,25 @@ func (u User) DeleteUser() revel.Result {
 	id := reqParam["id"]
 	revel.INFO.Println("DeleteUser id...====", id)
 
-	DB := dbConn()
+	if id != nil {
+		DB := dbConn()
 
-	// Modify some data in table.
-	rows, err := DB.Exec("DELETE FROM thx_employee WHERE id=?", id)
-	CheckErr(err)
+		// Modify some data in table.
+		rows, err := DB.Exec("DELETE FROM thx_employee WHERE id=?", id)
+		CheckErr(err)
 
-	rowCount, err := rows.RowsAffected()
-	revel.INFO.Println("Deleted %d row(s) of data.\n", rowCount)
-	revel.INFO.Println("Done.")
+		rowCount, err := rows.RowsAffected()
+		revel.INFO.Println("Deleted %d row(s) of data.\n", rowCount)
+		revel.INFO.Println("Done.")
 
-	if err != nil {
+		if err != nil {
+			result["auth"] = "fail"
+			revel.INFO.Println("DB Error", err)
+		}
+		result["auth"] = "success"
+	} else {
 		result["auth"] = "fail"
-		revel.INFO.Println("DB Error", err)
 	}
-	result["auth"] = "success"
 
 	return u.RenderJSON(result)
 }
@@ -154,29 +207,33 @@ func (u User) UpdateUser() revel.Result {
 	password := reqParam["password"]
 	role := reqParam["role"]
 	username := reqParam["username"]
-	revel.INFO.Println("UpdateUser id...====", id)
-	revel.INFO.Println("UpdateUser password...====", password)
-	revel.INFO.Println("UpdateUser role...====", role)
-	revel.INFO.Println("UpdateUser username...====", username)
+	//revel.INFO.Println("UpdateUser id...====", id)
+	//revel.INFO.Println("UpdateUser password...====", password)
+	//revel.INFO.Println("UpdateUser role...====", role)
+	//revel.INFO.Println("UpdateUser username...====", username)
 
-	DB := dbConn()
+	if id != nil && username != nil && password != nil && role != nil {
+		DB := dbConn()
 
-	// Modify some data in table.
-	rows, err := DB.Exec("UPDATE thx_employee SET id = ?, password = ?, role = ?, username = ?,  WHERE id = ?", id, password, role, username, id)
-	CheckErr(err)
-	rowCount, err := rows.RowsAffected()
-	fmt.Printf("Update %d row(s) of data.\n", rowCount)
-	fmt.Println("Done.")
+		// Modify some data in table.
+		rows, err := DB.Exec("UPDATE thx_employee SET password = ?, role = ?, username = ?  WHERE id = ?", password, role, username, id)
+		CheckErr(err)
+		rowCount, err := rows.RowsAffected()
+		fmt.Printf("Update %d row(s) of data.\n", rowCount)
+		fmt.Println("Done.")
 
-	if err != nil {
+		if err != nil {
+			result["auth"] = "fail"
+			revel.INFO.Println("DB Error", err)
+		}
+		result["auth"] = "success"
+	} else {
 		result["auth"] = "fail"
-		revel.INFO.Println("DB Error", err)
 	}
-	result["auth"] = "success"
 
 	return u.RenderJSON(result)
 }
 
-func main() {
-
-}
+//func main() {
+//
+//}
